+++ blog/resources/views/auth/register.blade.php
@@ -1,23 +1,35 @@
-@extends('layouts.app')
-
+@extends('layouts.auth.login')
+<title>Inscription - BLM</title>
 @section('content')
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-md-8">
-            <div class="card">
-                <div class="card-header">{{ __('Register') }}</div>
+{{-- arrow back to home page --}}
+<a href="{{ url('/') }}" class="arrow-west">
+    <span class="material-icons">
+        west
+    </span>
+</a>
+<div class="content d-flex justify-content-center align-items-center" style="height: 70vh; width:100%;">
+    <div class="container-fluid container-login" style="">
 
-                <div class="card-body">
-                    <form method="POST" action="{{ route('register') }}">
+        {{-- 2e container --}}
+        <div class="container d-flex justify-content-around align-items-center" style="">
+            <div class="card card-login d-flex flex-column justify-content-center" style="width: 40rem; height:auto">
+                <div class="card-body card-login-body">
+                    <form action="{{ route('register') }}" method="POST">
                         @csrf
-
-                        <div class="form-group row">
-                            <label for="name" class="col-md-4 col-form-label text-md-right">{{ __('Name') }}</label>
-
+                        <div class="form-row">
                             <div class="col-md-6">
-                                <input id="name" type="text" class="form-control @error('name') is-invalid @enderror" name="name" value="{{ old('name') }}" required autocomplete="name" autofocus>
-
-                                @error('name')
+                                <label for="firstname">Prénom<span style="color: red;">*</span></label>
+                                <input type="text" name="firstname" id="firstname" class="form-control form-control-lg @error('firstname') is-invalid @enderror" placeholder="Entrer votre prénom" required autocomplete="off" autofocus aria-autocomplete="none">
+                                @error('firstname')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                            <div class="col-md-6">
+                                <label for="lastname">Nom<span style="color: red;">*</span></label>
+                                <input type="text" name="lastname" id="lastname" class="form-control form-control-lg @error('lastname') is-invalid @enderror" placeholder="Entrer votre nom" required autocomplete="off" autofocus aria-autocomplete="none">
+                                @error('lastname')
                                     <span class="invalid-feedback" role="alert">
                                         <strong>{{ $message }}</strong>
                                     </span>
@@ -25,26 +37,19 @@
                             </div>
                         </div>
 
-                        <div class="form-group row">
-                            <label for="email" class="col-md-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
-
+                        <div class="form-row mt-3">
                             <div class="col-md-6">
-                                <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" value="{{ old('email') }}" required autocomplete="email">
-
+                                <label for="email">Email<span style="color: red;">*</span></label>
+                                <input type="email" name="email" id="email" class="form-control form-control-lg @error('email') is-invalid @enderror" placeholder="john@example.com" autocomplete="off" aria-autocomplete="none" required>
                                 @error('email')
                                     <span class="invalid-feedback" role="alert">
                                         <strong>{{ $message }}</strong>
                                     </span>
                                 @enderror
                             </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="password" class="col-md-4 col-form-label text-md-right">{{ __('Password') }}</label>
-
                             <div class="col-md-6">
-                                <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" required autocomplete="new-password">
-
+                                <label for="password">Mot de passe<span style="color: red;">*</span></label>
+                                <input type="password" name="password" id="password" class="form-control form-control-lg @error('password') is-invalid @enderror" placeholder="Entrer un mot de passe" autocomplete="off" aria-autocomplete="none" required>
                                 @error('password')
                                     <span class="invalid-feedback" role="alert">
                                         <strong>{{ $message }}</strong>
@@ -53,25 +58,38 @@
                             </div>
                         </div>
 
-                        <div class="form-group row">
-                            <label for="password-confirm" class="col-md-4 col-form-label text-md-right">{{ __('Confirm Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required autocomplete="new-password">
-                            </div>
+                        <div class="form-group mt-3">
+                            <label for="birth">Date de naissance</label>
+                            <input type="date" max="2002-01-31" class="form-control form-control-lg @error('birth') is-invalid @enderror" name="birth" placeholder="01/01/2000" autocomplete="off" aria-autocomplete="none">
+                            @error('birth')
+                                <span class="invalid-feedback" role="alert">
+                                    <strong>{{ $message }}</strong>
+                                </span>
+                            @enderror
                         </div>
-
-                        <div class="form-group row mb-0">
-                            <div class="col-md-6 offset-md-4">
-                                <button type="submit" class="btn btn-primary">
-                                    {{ __('Register') }}
+                        
+                        <div class="line"></div>
+                        <div class="create-account d-flex flex-column align-items-center justify-content-center">
+                            <div class="form-group">
+                                <button type="submit" class="btn btn-primary btn-lg btn-login">
+                                    Créer un compte
                                 </button>
+                            </div>
+                            <div class="form-group lost-password" style="text-align:center;">
+                                <a href="{{ route('login') }}">
+                                    Déja un compte ? Connectez vous ici 
+                                </a>
                             </div>
                         </div>
                     </form>
                 </div>
             </div>
+    
         </div>
     </div>
 </div>
+@endsection
+
+@section('scripts')
+
 @endsection